Rename main file to game logic and export it
